WIP commit - flags are now displaying on initial load with an initial name of the HTML tag
