+++ tob-web/src/app/timeline/slot-state.model.ts
@@ -1,0 +1,7 @@
+export class SlotState {
+  active: boolean = false;
+  focused: boolean = false;
+  hovered: boolean = false;
+  started: boolean = false;
+  ended: boolean = false;
+}
