+++ tob-web/tslint.json
@@ -9,7 +9,7 @@
       true,
       "check-space"
     ],
-    "curly": true,
+    "curly": [true, "ignore-same-line"],
     "eofline": true,
     "forin": true,
     "import-blacklist": [true],
